ocr-service/requirements.txt
paddlepaddle==3.0.0
paddleocr==2.9.1
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.18
Pillow==11.1.0
